FocusPreservationTests.ChangingFocusedNodeResetsFocusPreservationState triggers a debug assertion https://bugs.webkit.org/show_bug.cgi?id=202145 <rdar://problem/51273128>
Reviewed by Tim Horton.

Source/WebKit:

Fixes an assertion in -tableView:cellForRowAtIndexPath: to only require the cell to have a non-zero
width in the case where the table view itself has a non-empty width. When run under TestWebKitAPI,
this assertion currently fires because the view controller used to show the popover isn't actually
presented, so the view controller's view (i.e. the UITableView) ends up being empty. This causes all
of its table view cells to subsequently become empty upon reuse.

* UIProcess/ios/forms/WKFormSelectPopover.mm:
(-[WKSelectTableViewController tableView:cellForRowAtIndexPath:]):

Tools:

Fixes this API test to work on iPad as well as iPhone. Importantly, we show a popover when focusing
select elements on iPad, which means that there is no -inputView present on WKContentView, and upon
dismissing the form accessory view, the select element is still focused.

* TestWebKitAPI/Tests/ios/FocusPreservationTests.mm:


Identifier: 215752@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250306 268f45cc-cd09-0410-ab3c-d52691b4dbfc