[Web Animations] Removing an element should only cancel its declarative animations https://bugs.webkit.org/show_bug.cgi?id=202165
Reviewed by Dean Jackson.

LayoutTests/imported/w3c:

Mark a WPT progression and a change in FAIL expectation in a test where we used to cancel an animation and now don't.

* web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-context-expected.txt:
* web-platform-tests/web-animations/timing-model/animations/canceling-an-animation-expected.txt:

Source/WebCore:

When an element is removed, an Animation created using the JS API should not be canceled.
Only declarative animations should be.

* animation/AnimationTimeline.cpp:
(WebCore::AnimationTimeline::elementWasRemoved):

Identifier: 215776@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250335 268f45cc-cd09-0410-ab3c-d52691b4dbfc