TestWebKitAPI.WebKit.ReloadWithDifferingInitialScale is failing consistently https://bugs.webkit.org/show_bug.cgi?id=202273
Reviewed by Aakash Jain.

The content width scaling limits added in r250361 prevent this test from scaling down the content width (2000px
wide) to fit the viewport (375px). To avoid this issue, change the content width from 2000px to 1500px, which is
still a very wide content width for a web page.

* TestWebKitAPI/Tests/WebKitCocoa/ReloadWithDifferingInitialScale.mm:


Identifier: 215824@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250391 268f45cc-cd09-0410-ab3c-d52691b4dbfc