Flaky Test: inspector/canvas/updateShader.html https://bugs.webkit.org/show_bug.cgi?id=202186 <rdar://problem/55716053>
Reviewed by Joseph Pecoraro.

If the `WebGLProgram` outlives it's `WebGLRenderingContext`, the `ScriptExecutionContext*`
that was provided in the constructor won't be invalidated leading to the bad access crash.

Rather than pass the `ScriptExecutionContext*` directly, have `WebGLProgram` inherit from
`ContextDestructionObserver` so that it can propertly invalidate (and notify Web Inspector)
when the related context is about to be destroyed.

Test: inspector/canvas/updateShader.html

* html/canvas/WebGLProgram.h:
(WebCore::WebGLProgram::scriptExecutionContext const): Deleted.
* html/canvas/WebGLProgram.cpp:
(WebCore::WebGLProgram::WebGLProgram):
(WebCore::WebGLProgram::contextDestroyed): Added.


Identifier: 215863@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250434 268f45cc-cd09-0410-ab3c-d52691b4dbfc