[JSC] AI folds CompareEq wrongly when it sees proven Boolean and Number https://bugs.webkit.org/show_bug.cgi?id=202382 <rdar://problem/52669112>
Reviewed by Saam Barati.

JSTests:

* stress/compare-eq-bool-number-folding.js: Added.
(test):

Source/JavaScriptCore:

If CompareEq(Untyped, Untyped) finds that it gets proven Boolean and Number types on its arguments,
we fold it to constant False. But this is wrong since `false == 0` is true in JS.
This patch adds leastUpperBoundOfEquivalentSpeculations, which merges Number, BigInt, and Boolean types
if one of them are seen.

* bytecode/SpeculatedType.cpp:
(JSC::leastUpperBoundOfEquivalentSpeculations):
(JSC::valuesCouldBeEqual):

Identifier: 215913@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250536 268f45cc-cd09-0410-ab3c-d52691b4dbfc