Crash in debug build with imported/w3c/web-platform-tests/web-animations/timing-model/timelines/update-and-send-events-replacement.html https://bugs.webkit.org/show_bug.cgi?id=202583
Patch by Antoine Quint <graouts@apple.com> on 2019-10-04
Reviewed by Dean Jackson.

Ensure we remove the animation from the timeline it's currently associated with since it may have changed since its creation. This caused
a crash because we have an assertion in AnimationTimeline::removeAnimation() to check that the animation belonged to the timeline it's called on.

* animation/DocumentTimeline.cpp:
(WebCore::DocumentTimeline::removeReplacedAnimations):

Identifier: 216071@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250737 268f45cc-cd09-0410-ab3c-d52691b4dbfc