[macOS WK2 Debug ] Flaky Test: animations/play-state-in-shorthand.html https://bugs.webkit.org/show_bug.cgi?id=192501
Patch by Antoine Quint <graouts@apple.com> on 2019-10-07
Reviewed by Simon Fraser.

This test used to rely on timers to check the computed style at various animation times and also to check the "animation-play-state"
to "paused" via the "animation" shorthand CSS property. This was flaky by design since timers could not fire with high enough fidelity
to ensure the animation time matched the expectation. With the availability of the Web Animations API, we can simply use the "playState"
property on the Animation object reflecting the CSS animation to check that the change is accounted for.

* animations/play-state-in-shorthand-expected.txt:
* animations/play-state-in-shorthand.html:
* platform/ios-simulator-wk2/TestExpectations:
* platform/mac/TestExpectations:

Identifier: 216110@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250781 268f45cc-cd09-0410-ab3c-d52691b4dbfc