[MSE][GStreamer] Fix video sometimes not appearing https://bugs.webkit.org/show_bug.cgi?id=201401
Reviewed by Xabier Rodriguez-Calvar.

LayoutTests/imported/w3c:

Added a test that reproduces the sequence of operations that was
causing the video to not appear in WebKitGTK and checks that this time
there is visible output.

* web-platform-tests/lint.whitelist:
* web-platform-tests/media-source/mediasource-video-is-visible-expected.html: Added.
* web-platform-tests/media-source/mediasource-video-is-visible.html: Added.
* web-platform-tests/media-source/mp4/test-a-1s.mp4: Added.
* web-platform-tests/media-source/mp4/test-a-1s.mp4-manifest.json: Added.
* web-platform-tests/media-source/mp4/test-v-1s-blue.mp4: Added.
* web-platform-tests/media-source/mp4/test-v-1s-blue.mp4-manifest.json: Added.
* web-platform-tests/media-source/webm/test-a-1s.webm: Added.
* web-platform-tests/media-source/webm/test-a-1s.webm-manifest.json: Added.
* web-platform-tests/media-source/webm/test-v-1s-blue.webm: Added.
* web-platform-tests/media-source/webm/test-v-1s-blue.webm-manifest.json: Added.

Source/WebCore:

The code in MediaPlayerPrivateGStreamer::changePipelineState() was
supposed to call `ensureGLVideoSinkContext()` on upwards transitions
to PAUSED but the code did not take into account non-step-by-step
state transitions, which happens frequently with playbin3 in the MSE
case.

Before the patch, when playbin3 transitioned from READY to PLAYING
without stopping for preroll this call would not be made and the
texture IDs received at the sink would not correspond to the
compositor GL context, leading to artifacts (often the player controls
or a blank screen).

Test: imported/w3c/web-platform-tests/media-source/mediasource-video-is-visible.html

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::changePipelineState):

LayoutTests:

Added an exact expectation picture to avoid errors in the test runner
(only a fuzzy match is expected, and different platforms render video
with slightly different colors).

* platform/mac/imported/w3c/web-platform-tests/media-source/mediasource-video-is-visible-expected.html: Added.
* platform/mac/TestExpectations:


Identifier: 216219@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@250922 268f45cc-cd09-0410-ab3c-d52691b4dbfc